in progress - documents zip file management
git-svn-id: https://svn.code.sf.net/p/openuss/code/trunk/openuss-plexus@319 60046477-0c10-0410-bf4b-e552a4ff1768

+++ plexus/plexus-web/src/main/java/org/openuss/web/repository/RepositoryServlet.java
@@ -85,7 +85,7 @@ public class RepositoryServlet extends HttpServlet {
 		}
 		response.setContentType(file.getContentType());
 //		response.setContentType("application/octet-stream");
-		response.setContentLength(file.getFileSize());
+		response.setContentLength(file.getFileSize().intValue());
 	}
 
 	private void sendFileNotFound(final HttpServletResponse response) {
